backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-litellm>=0.4.0",
    "langchain-openai>=1.1.8",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-postgres>=3.0.4",
    "netra-sdk>=0.1.67",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.1",
]
